don't use Subscription This was meant to have been committed in 7b45827468ddc1354eecadde1074c6773f535f02, but for some reason that didn't happen.
The Subscription class is too inflexible, so we use a plain
BasicQueuingConsumer instead (and, furthermore, hide it from the
interface).

As a result there is less "magic" happening behind the scenes, and the
new SetupDelegate provides full control over resource
declarations. The downside is that no resource declaration is
performed at all by default, but that's because there are just too many
perfectly sensible ways in which the resources could be set up, and it
doesn't make sense to choose one of them as the default over the
others.